# swagger.yaml
swagger: "2.0"
info:
  title: Dog Cat Matcher API
  description: Test server of Dog Cat Matcher project
  version: 0.1.0

basePath: /api
schemes:
  - http

paths:

  /animals:
    get:
      summary: Return a list of all animals
      responses:
        200:
          description: A list of animal objects
          schema:
            type: array
            items:
              $ref: "#/definitions/Animal"

  /animal/{id}:
    get:
      summary: Return an animal object by id
      parameters:
        - in: path
          name: id
          type: string
          required: true
          description: animalId
      responses:
        200:
          description: An animal object
          schema:
            $ref: "#/definitions/Animal"
        401:
          description: Unauthorized
        404:
          description: Animal with specified ID was not found
        default:
          description: Unexpected error

  /animal:
    post:
      summary: Add a new animal
      parameters:
        - in: body
          name: animal
          description: An animal object
          schema:
            type: object
            properties:
              name:
                type: string
              description:
                type: string
              tags:
                type: array
                items:
                  type: string
      responses:
        200:
          description: Created
          schema:
            type: object
            properties:
              id:
                type: string
        401:
          description: Unauthorized
        default:
          description: Unexpected error

    put:
      summary: Update an animal object
      parameters:
        - in: body
          name: animal
          description: An animal object
          schema:
            $ref: "#/definitions/Animal"
      responses:
        200:
          description: Success
        401:
          description: Unauthorized
        404:
          description: Animal with specified ID was not found
        default:
          description: Unexpected error
    
    delete:
      summary: Delete an animal object
      parameters:
        - in: body
          name: animalId
          description: An animal id
          schema:
            type: object
            properties:
              id:
                type: string
      responses:
        200:
          description: Success
        401:
          description: Unauthorized
        404:
          description: Animal with specified ID was not found
        default:
          description: Unexpected error  

  /uploadPhoto:
    post:
      summary: Upload an animal photo
      consumes:
        - multipart/form-data
      parameters:
        - in: formData
          name: photo
          type: file
          description: A photo to upload
          required: true
        - in: formData
          name: id
          type: string
          description: Animal ID to upload to
          required: true
      responses:
        201:
          description: Success
        400:
          description: No photo found
        401:
          description: Unauthorized
        404:
          description: Animal with specified ID was not found
        default:
          description: Unexpected error

definitions:
  Animal:
    type: object
    required:
      - id
    properties:
      id:
        type: string
      name:
        type: string
      description:
        type: string
      photoUrl:
        type: array
        items:
          type: string
      tags:
        type: array
        items:
          type: string